Add a seed project for creating vCD UI plugins.
The seed project uses versions of Clarity and Angular that make it
compatible with the 9.1.0 release of vCloud Director.

The seed project can be built and registered as-is with a vCD instance,
or it can be customized to provide functionality that is actually
useful.  Examples of managing styles, internationalized text content,
and image assets are included in this sample.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>