Add simplified component and module as plugin.
This change collapses the route definition into the module, and includes
Clarity for use by components.  The component also adds the
content-container class to the component selector to ensure more proper
layout of the plugin with respect to its container.

Aside from some cosmetic changes, this plugin is identical to, and
supersedes, the original stub module and component.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>