Add module supporting sub-navigation as plugin.
This changes introduces a module with a routing pattern that supports
managing parent/child routes within the plugin.

The main component (SubnavComponent) creates a vertical navigation based
container that is consistent with Clarity standards and with the look
and feel of sub-navigation that exists in the vCD 9.1.0 user interfaces.
It allows for easy additions of single level sub-navigation via the
`navItems` property.

This specific implementation adds 3 child routes: one to a simple text
page, one to a page with an image asset, and one to an IFrame that
renders a page from an external site.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>