Move JSON xjc plugin to separate project.
A new project, vcd-xjc-plugins, will contain customizations for XML to
Java compilation, allowing those customizations to be shared by multiple
bindings project.  The main driver for this is to make the JSON subtype
generation usable by both extensibility and REST APIs.

A side effect of this change is that the maven-compiler plugin doesn't
need to execute as a different phase, and the xjc goal can be processed
during the more typical generate-sources phase.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>