Fix extensibility-api-bindings build. (#14)
* Allow extensibility-api-bindings to build.

The com.sun.tools.xjc.Plugin resource was missing, so the xjc plugin
wasn't able to process the -Xjson-api extension correctly.

The namespace is also changed for the JsonMappingInfoPlugin to avoid
future collision with other custom plugins.

* Move JSON xjc plugin to separate project.

A new project, vcd-xjc-plugins, will contain customizations for XML to
Java compilation, allowing those customizations to be shared by multiple
bindings project.  The main driver for this is to make the JSON subtype
generation usable by both extensibility and REST APIs.

A side effect of this change is that the maven-compiler plugin doesn't
need to execute as a different phase, and the xjc goal can be processed
during the more typical generate-sources phase.

* Move REST version plugin to common project.

This moves the RestApiVersionsPlugin and supporting classes to the
vcd-xjc-plugins project.  Maven pluginManagement is used to ensure that
all xjc processing is using consistent versions of plugins and
dependencies.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>