Add Typescript binding generation for XML APIs.
This change adds a bindings generator that processes the JAXB
annotations from vCD's Java API bindings and uses that information to
construct Typescript bindings of the XSD-based schemas.  A Maven plugin
is also included.

This tool focuses specifically on the XML-based APIs because the newer
APIs available at the /cloudapi endpoint are based on YAML definitions
and have generation support available through OpenAPI tooling.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>