Add new overwrite type to bindings generation.
The overwrite flag was previously a boolean.  It has been changed to an
enum to support full rewrite, full failure, or a merge attempt.  The
merge is useful for incorporating Open API models which have their own
generation mechanism, and allowing the barrels to take the Open API
models into account.

The api-client project has been updated to include Open API bindings.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>