Change scope functionality
Rename change scope to change org scope

The name of component and services has to match with thier action.

Implement change scope func

Create new components and services - allows you to change plugin scope.

Add choose scope component to upload component

User have to select the scope of the plugin while uploading.

Implement change scope validation.

Validation is needed to prevent unnecessary requests.
Alert component is added in change scope component to show message.

Add refresh action on change scope close

The list of the plugins has to be in sync after users action.

Add conditional refresh of plugin list

Datagrid data has to be refreshed only if action is excuted.

Change org service subject to behavior

Get last emitted value on subscription.

Implement unpublish func and tracking

Break plugin manager to smaller services

This is for better code readability. (Facade Pattern)

Implement middleware promise flow control

This is needed because we want to track the progress of the change req.

This is for better visablility of the request list.
This changes are applied to upload and change scope plugin components.

Create pop up for the request tracker

This is for better visablility of the request list.
This changes are applied to upload and change scope plugin components.

Create error notifyer component

Will be used to show error into the screen.

Handler errors across the components

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Add open error notifyer property

It is used to open and close the error notifyer.

Add error handling of sendzip method

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>