Integrate the @vcd/sdk package into the seed project.
This change includes the @vcd/sdk and @vcd/bindings packages into the
seed project, and specifies them for inclusion in the rollup process.

The SimpleModule has been updated to import the VcdSdkModule and use the
VcdApiClient as a provider.  The module initialization will set the
authentication on the client, allowing the SimpleComponent to display
the logged in user and tenant.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>