[sdk] Place container hooks into SDK
The upcoming version of VCD places the container hooks into the SystemJS registry
under @vcd/common.

This adds support to the SDK to import and re-export these container hooks.
This means that consumers only have to import the SDK, and can access the hooks
as if they were part of the SDK.

The plugin seed has therefore been updated to consume the SDK and use its exports, rather than
the old @vcd-ui/common or the new @vcd/common. Running its build works - it has been tested
against a new build of the VCD UI and it loaded perfectly.

Signed-off-by: David Byard <dbyard@vmware.com>