Move container hooks into @vcd/sdk package (#24)
Instead of each plugin having a copy of the definitions for vCD integration, the definitions are now provided by the @vcd/sdk package.

The definitions are imported from the @vcd/common module of the SystemJS registry of the vCloud Director UI.  This change also re-exports those definitions so that they are available directly to plugins that have already integrated with the @vcd/sdk package.

Note: previous versions of the vCloud Director UI imported the hooks from @vcd-ui/common.  The import/re-export flow in this commit will need to be changed at some point to provide backward compatibility.

Testing Done:
The plugin seed has therefore been updated to consume the SDK and use its exports, rather than
the old @vcd-ui/common or the new @vcd/common. Running its build works - it has been tested
against a new build of the VCD UI and it loaded perfectly.

Signed-off-by: David Byard <dbyard@vmware.com>