Move SDK toward better backward compatibility. (#25)
Note: this change does not yet make the SDK fully backward compatible to
the 9.1 release of vCloud Director, but it is a step in the right
direction.  This change will attempt to register the container hooks
from @vcd/common which is a new location for hooks.  If that fails it
will attempt to load hooks from @vcd-ui/common (the old location) before
failing.

This change also makes some improvements to the SDK and to the seed
project.  Since the container hooks are now part of the @vcd/sdk
package, the VcdApiClient can manage its own injection of API base URL,
authentication token, etc. and so the plugin module no longer needs to
include this plumbing.  A side effect is that the clients session info
is not immeidately available, so the client methods that parsed specific
fields from the session have been changed to return Observables tracking
a BehaviorSubject.

Testing Done:
Built and deployed the seed plugin to a vCD instance and verified the
plugin successfully loaded.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>