Add localization capabilities to the @vcd/sdk. (#27)
Previously, plugins relied on container hooks to the I18nModule to
perform string translations.  This meant that all the translation keys
for a plugin were loaded into a shared translation service in the core
vCloud Director UI.  This had the unfortunate side effect of plugins
being able to overwrite each other's translation keys.  This change adds
a translation pipe and service to the SDK, where it can be used as a
scoped provider for the individual plugins.

This initial implementation handles basic positional interpolation, but
not the full ICU message format.  It supports either loading all
translations from a single i18n.json file (the default) or loading
individual language files: en.json, de.json, etc.

A plugin module was also added to reduce some of the boilerplate needed
to create the individual plugin modules.

The seed project was updated to use the new translation service and
PluginModule instead of the I18nModule.

Existing plugins that upgrade to this version of the SDK will need to
remove references to I18nModule.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>