Add multisite and paging capabilities to query API.
This change adds the ability to specify multisite fanout for queries,
either as a global fanout, or targeted to specific locations.

The change also adds the ability to navigate to additional pages in a
query result, using the navigation rels for first, previous, next, and
last.  This style of paging is compatible with both single site and
multisite queries.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>