Add basic CRUD capabilities to @vcd/sdk. (#30)
This change adds both synchronous and asynchronous versions of create,
update, and delete to the VcdApiClient, along with a get.  Al of these
methods are to be considered low level API calls-they rely on pure
relative URL endpoints rather than the link navigation that is typical
of the core vCloud Director API.  These calls are ultimately a necessary
part of the SDK, since the SDK needs to support API extensions in
addition to the core API.  These low level calls WILL give a client the
capability to consume most of the existing APIs.

All ****Async versions of the functions return a TaskType.  The task is
parsed from either the Location header (OpenAPI style) or the response
body (XSD-based API style) as appropriate.

Two convenience methods have also been added: getEntity() and
removeItem().  getEntity() resolves a URN-based id or an entity
reference to a specific instance of an EntityReferenceType by using the
entityResolver URL and then automatically following the response links.
removeItem() removes a ResourceType by treating it as a navigable
object, ensuring that the appropriate 'remove' link exists, and calling
the appropriate href, all without the API consumer needing to specify
any URL fragments.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>