Fix API calls to work on the /cloudapi endpoint. (#31)
Content-Type header for /api endpoints must match application/*+json to
be correctly evaluated by @Consumer annotations in the API handlers.
For the /cloudapi space, that header must be application/json.  Multiple
content types is not an option, so this fix simply uses one or the other
based on the request URL containing the 'cloudapi' string.

It isn't elegant.  It works for now.

Testing Done:
Performed a GET and PUT on the /cloudapi/branding endpoint to update the
portal name, and did a GET on /api/session and several query endpoints.
Verified that no calls returned a 415 error.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>