Merge requests
Merge operation is used to excute requests
with single subscription and independantlly
of each other.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Unsubscrobe from modal opening

Every subscription has to closed because can cause memory leaks.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Register plugin service

This service uses vcd api client.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Revert "Register plugin service"

This reverts commit 7aeecbc6492efaf8fbec32d22ceb8664d7c05e52.

Configure rollup.config file

The configuration wasn't for this version so
some properies had to be renamed.
@vcd-ui/common is also added in external
list of dependencies.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Remove httpClient import from subnav module

HttpClient service is not used so it has to be removed.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Replace current plugin interfaces with these from bindings

There is duplication in plugin-lifecycle interfaces
and these in bindigs. So these from plugin-lifecycle
is replaced with this in the bindings to avoid
dupliations.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Start use PluginDesc instead of uipluginmetadata

This interface is not compatible with
this on the server-side. Some property
names are incorrect and request fails with
status 500.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Replace Plugin interface with UiPluginMetadataResponse

Plugin interface doesn't exist and
UiPluginMetadataResponse is used
instead.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Make id of ChangeScopePlugin interface not required

The id has to not requirde because
sometimes lika e parameter is passed array
of UiPluginMetadataResponse objects where
id is not required so the two interfaces has
to be with identical id property.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>