Fix serialization issues with underscores in models.
The Swagger codegen for Typescript defaults to converting the Swagger
definition fields to camelCase instead of leaving them in their original
style.  This is a bit odd, since it causes JSON.stringify to produce
invalid JSON for the model.  The codegen process has been modified to
force use of original field name.  The codegen process also now includes
the generated typescript dir in a mvn clean.

Also fixed the string const definitions in field name generation, and
converted the @vcd/bindings package to an es6 module instead of a
commonjs module.  The rollup of plugins was having some issues trying to
determine the named exports from this package, causing the
rollup.config.js to require undeterministic tweaks.

Testing Done:
Regenerated models.  Visually inspected UiPluginMetadata to ensure
property was generated as tenant_scoped and not tenantScoped.  Ensured
/index in barrels, and successful build of @vcd/bindings.  Verified
successful build of a plugin that used bindings that were previously
problematic.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>