Add styles for status component
The small separator of the column of the
datagrid was overlaping with the dropdown
menu of scope action, so z index is added
on the btn action bar.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Handle the case when change org scope
isn't tracked by tracker service

There was no logic which handles
the case when the user don't want
track the requests. So this logic was
implemented.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>