Upgrade change tenant scope and upload workflows
New method is added in plugin service which
gets all tenants, for which the plugin is published.
Related with this new request, new interface is
created to describe the response.

The previous version changing the tenant scope
functionality wasn't the easiest way to achieve that.
So the publish and unpublish for all and specific
tenants is moved to common modal.
The current workflow is: the first radio
button unpublish for all tenants, second publish for
all tenants and the third, the datagrid, publish and
unpublish for specific tenants, with selection
and deselection.

We have this changes in upload wizard too, plus
the page where you can select for which tenants you
want to publish the plugin, is now conditional,
so the page will be shown if the plugin is scoped
for tenants.
Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>