Add filtering and field selection to Query.Builder.
This change adds the ability to specify which fields of a query record
to populate.  If specific fields are not requested (the default
behavior), the entire record is returned.

A Filter.Builder has also been added to support programmatic creation of
filter conditions, using a fluent API style heavily influenced by the
Apache CXF SearchConditionBuilder.  The builder supports creating
conditions for equals, not equals, less than, less than or equal,
greater than, greater than or equal, and wildcard equals (like), as well
as conjunctions and disjunctions.  Strings, numbers, and booleans are
currently supported where appropriate by the different comparison
operations.

Testing Done:

Executed the following statement in a plugin:
let builder = new Filter.Builder();
let filter = builder.is('isEnabled').equalTo(true).and().or(
    builder.is('name').like('%aaaa', Filter.MatchMode.START),
    builder.is('name).like('org', Filter.MatchMode.START)
).query();
this.client.query(Query.Builder.ofType('organization').links(false)
    .filter(filter).fields('displayName')).subscribe(console.log);

Verified that the URL was built with a valid FIQL string, that the
query did the appropriate level of filtering, and that only the
displayName was populated in the respose records.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>