Add lifecycle management UI for plugins (#33)
This change adds a plugin that can be used to upload and manage other UI plugins.

The plugin is intended to be consumed from the provider portal.  It allows users to select ZIP files of plugins from their local file system and upload them into vCloud Director.  The user can then perform additional management tasks for the plugins including, scoping access to the provider or tenant portal, publishing the plugin to specific tenants, enabling/disabling the plugin, and removing the plugin.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>