Publish api packages as es2015 modules with es5 target.
This addresses an issue where plugins failed to load in IE11 and is just
the more compatibile format in general with the current state of
browsers.

This fixes Issue #37.

Updated plugin-lifecycle plugin to use new module which required a small
change to the rollup config to suppress "`this` has been rewritten to
`undefined`" warnings.

The vcd-plugin-seed had a slightly larger change.  In addition to the
warning suppression, the commonjs plugin was removed to ensure pure
es2015 modules as the starting point, and the typescript plugin was
changed to rollup-plugin-typescript2, which should do a better job of
syntactic and semantic diagnostic messages at build time with a nominal
hit to build speed.

Testing Done:
Deleted node_modules for api-client packages and ran `yarn bootstrap`.
Deleted node_modlues of plugin-lifecycle and vcd-plugin seed before
executing `yarn install && yarn build`.  Uploaded both plugins to a vCD
cell and verified that they correctly loaded in IE11, and continued to
load in Chrome.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>