Support sorting in Query.Builder.
This change adds the ability to specify 0 or more sort parameters to
apply to a query.  A sort parameter is the field name to sort, and an
optional directive to reverse the sort order (an ascending sort is the
default). When multiple sort parameters are specified, they are applied
in the order that they are specified.  This change addresses Issue #40.

Also included in this change is a fix for Issue #42. The combination of
rest parameters and function overloads with different return types was
resulting in the wrong overload being called at runtime, which caused
subsequent calls to fail.  Removing the rest parameter in favor of an
optional array param is a bit more verbose for the caller, but seems to
be handled correctly.

Testing Done:
Built the following filter:
    let filter = builder.or(
        builder.is(QueryResultOrgRecordType.Fields.NAME).like('aaaa', Filter.MatchMode.START),
        builder.is(QueryResultOrgRecordType.Fields.NAME).like('org', Filter.MatchMode.START)
    ).query();

Verified that a valid query string was constructed:
    type=organization&format=idrecords&links=false&pageSize=25&fields=name,displayName&filter=(name==aaaa*,name==org*)

Also built the following filter:
    let filter =
    builder.is(QueryResultOrgRecordType.Fields.IS_ENABLED).equalTo(true).and()
        .or(
            builder.is(QueryResultOrgRecordType.Fields.NAME).like('aaaa', Filter.MatchMode.START),
            builder.is(QueryResultOrgRecordType.Fields.NAME).like('org', Filter.MatchMode.START)
    ).query();

Verified that a valid query string was constructed:
    type=organization&format=idrecords&links=false&pageSize=25&fields=name,displayName&filter=isEnabled==true;(name==aaaa*,name==org*)

Built queries with 0, 1, and 2 sort params (a mix of default and reverse
sorting). Verified that the correct query string was produced in each
case and that the API call completed successfully.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>