[housekeeping] Scoped HTTP client in extension SDK
During research it was discovered that the currently-configured HttpClient is a singleton one, meaning
that every single extension module and the container all use the same instance.  This could get
hairy very quickly, as interceptors would also be global.

This subclass `HttpClient` as `VcdHttpClient`.  Within the extension and SDK, users can inject the VcdHttpClient,
and because it is compiled in the module in will always rehave a unique reference per-extension and therefore
provide the requisite module-level scoping.

Testing done: Built the test dashboard extension using this, and it still works properly.
Signed-off-by: David Byard <dbyard@vmware.com>