[ui] Fix re-export of VcdApiClient and ApiResultService
The VcdApiClient change only was tested with rollup.js
consumers.  Webpack builds tests types a lot more strongly, and
exposed that the re-export of VcdApiClient/ApiResultService
was broken.

The clauses have been updated to generate correct code.

Testing done: Testing with the plugin seed, as that has a webpack build.
Signed-off-by: David Byard <dbyard@vmware.com>