[ui] SDK transfer client
Add support for file transfers in the SDK. Implemented using
a pure Observable flow, with the following features:

    File and Blob can be uploaded
    Authentication is shared with the VcdApiClient
    Progress can be monitored bypassing an Observable
    Each chunk will be a maximum of 50MiB
    Each chunk will be retried 5 times

To initiate a transfer, use the VcdApiClient.startTransfer method.
This will return a VcdTransferClient instance. This can be
then provided a file and a progress callback via its upload method.

Testing done: A custom, modified version of the plugin lifecycle manager was created using this. While it is not ready for commit, it successfully uploaded files using multiple chunks and single chunks.
Signed-off-by: David Byard <dbyard@vmware.com>