[ui] Support header HATEOAS links as if they were body links
Add an interceptor that converts HATEOAS header links as if they were links on the
response object, as long as the response is either `null` or an object.

If the response object is null, it will become an object of type `{link: LinkType[]}`.

If the response object is not null and has a `link` array, the header links will be
combined with the existing links.

If the response object is not null and has no `link` array, the header links will
be added as a `link` field.

This means all links should always be available in `response.body.link`, and of type
LinkType[].

Testing done: Manual debug breakpoints on various API calls made by a test plugin.
Signed-off-by: David Byard <dbyard@vmware.com>