[ui] SDK transfer client
Add support for file transfers in the SDK. Implemented using
a pure Observable flow, with the following features:
 * File and Blob can be uploaded
 * Authentication is shared with the VcdApiClient
 * Progress can be monitored bypassing an Observable
 * Each chunk will be a maximum of 50MiB
 * Each chunk will be retried 5 times

To initiate a transfer, use the VcdApiClient.startTransfer method.
This will return a VcdTransferClient instance. This can be then provided
a file and a progress callback via its upload method:
```
   const transferClient = this.vcdApiClient.startTransfer(postEndpoint, payload);
   transferClient
       .upload(myFile, this.progressSubject)
       .toPromise()
       .then(() => this.finished = true, (err) => this.error = err.message);
```

The following unit tests are provided:
 - upload a file in a single chunk
 - upload a file in multiple chunks
 - receiving sensible progress during a chunked upload
 - handling an unchunked upload failing with a 404
 - handling a chunked upload failing with 404
 - handling an unchunked upload failing with a 500
 - handling a chunked upload failing with 500
 - handling an unchunked upload failing with a read error
 - handling a chunked upload failing with a read error
 - hadling a chunked upload failing with intermittent 500 errors
   (retry logic)

Testing done: A custom, modified version of the plugin lifecycle manager was created using this. While it is not ready for commit, it successfully uploaded files using multiple chunks and single chunks; unit tests
Signed-off-by: David Byard <dbyard@vmware.com>