[ui] Support entity action extension points.
Export the EntityActionExtensionComponent base class needed for creating
entity action extension points.

This is an abstract class exported by SystemJS from the container.  This
caused a right headache when defining typing - it seems you can't easily
type an arbitrary JS object as an abstract class.  Even using `declare`
on the abstract class type causes compilation errors.

The used approach does pollute the namespace a bit with an underscored
variant, but it ensures that if you `extend` the version without the
underscore, typescript enforces the abstract methods.

Testing done: modified the plugin seed to ensure the typing worked as expected.
Signed-off-by: David Byard <dbyard@vmware.com>