Support negotiating API version with a vCD backend.
This change specifies a set of API versions that the SDK is compatible
with, ordered in preference from newest to oldest. When no explicit API
version is specified, the VcdApiClient will query vCloud Director for the
set of API versions that it supports, and attempt to match a server
version to an SDK version in order of preference.

If an explicit version is specified, this negotiation does not happen.

Because a valid version is required for most API calls, including getting
the current session, the setAuthentication() method subscription has been
moved out of the constructor. It was not really a great design to have an
async operation as part of the constructor that then was a pre-requisite
to executing any other methods. To successfully move the session retrieval
out of the constructor, to make it rely on possible content negotiation,
and to preserve backward compatibility for existing consumers of the SDK,
a "run once" async chain has been added to the beginning of all of the
fundamental public-facing methods.

Testing done:
Successful execution of unit tests. Additional testing currently in
progress.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>