Normalize JSON serialization differences in 9.1. (#75)
This change addresses a change in JSON serialization that happened
between vCD 9.1 and vCD 9.5. The JSON mapper used in 9.1 presented some
additional metadata and put the actual response payload inside a 'value'
property. Post-9.1 vCloud Director returns the response payload directly
as the response body.

The handling is done at the HttpClient level, in an interceptor, to
allow the object bindings that happen later in the chain to just work.

Testing done:
Updated the seed plugin to use this new SDK build. Created a
distribution of the plugin and deployed to a 9.1 testbed and a 9.5
testbed. Verified that the plugin loaded and displayed the session
information in both testbeds. Inspected the Network tab in Chrome to
verify that the responses were returned in 2 different formats (wrapped
and unwrapped).

Added query call to seed project and verified that `record` payload was
correctly processed in both a 9.1 and 9.5 testbed.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>