Fix build process for publishing @vcd/sdk package.
This change introduces the frontend-maven-plugin into the api-client
project. This will allow an install/test/deploy lifecycle in TravisCI
that is compatible both with the api-client Java project, and with NPM
packages in our other GitHub repositories.

The structure of api-client is currently more nested than it needs to
be. That will be addressed in a follow-up change once the @vcd/sdk
package is available through NPM and not referenced as a local file path
in a consumer's package.json.

The Maven process writes the build artifacts to a target/ director, in
line with Maven standards and to allow the non-Maven build processes to
continue to work as-is.

Testing done:
- Built the api-client SDK with `mvn install` and verified output in the
target directory.
- Deployed the api-client SDK with `mvn deploy -Prelease and verified
output of the --dry-run matched the dist/ content
- executed `yarn` and `yarn bootstrap` directly and verified that the
package still built and copied to build/ and dist/, not nested inside
the target/ directory.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>