Re-add container hook for session org id.
This token used to be available in a very early version of the SDK, and
it has been available from the host since 9.1. Adding it back to
container hooks by request, and because it will be useful to drive
act-as functionality.

Testing Done:
Injected new const into component of seed plugin. Wrote value out to
console and verified it was available and correct.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>