Support act-as functionality in API client.
The 'act-as' functionality was introduced to API calls in version 9.1 of
vCloud Director. It allows a provider to make API requests in the
context of a specific tenant. The biggest use cases for this
functionality is to allow a provider access to a tenant's portal, while
showing only the entities that are visible to the tenant, instead of all
the entities visible to the provider.

The feature is enabled automatically when a provider navigates to the UI
portal for a tenant, as determined by the container hook for
SESSION_SCOPE and the organization of the looged in user's session. The
feature can additionally be manually enabled/disabled.

Testing Done:
Executed new unit tests to verify enablement, disablement, and automatic
enablement of the act-as feature.

Ran a seed plugin and verified in the Chrome Network tab that the
request header `X-VMWARE-VCLOUD-TENANT-CONTEXT` appeared as appropriate.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>