Clean up the build and CI process for @vcd/sdk.
The Maven-based build process for the @vcd/sdk was an artifact of the
previous build process which required Java libraries to be built in
order to build the Typescript bindings for the SDK. Those processes are
no longer necessary, as the schemas and bindings are all managed in a
different repository and available through public Maven repositories.

Additionally, yarn was removed in favor of npm to standardize on the
recommended tooling in other Typescript libraries like the seed plugin.

The Travis configuration was updated to use a matrix so that Java things
can build like Java things, and Node-y things can build like Node-y
things.

Testing Done:
Executed the new NPM scripts locally for build, test, and deploy and
verified that they still work as expected. Copied the produced @vcd/sdk
artifacts to a seed plugin, built and deployed the plugin to a vCD
instance, and verified that the plugin loaded without errors.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>