Modify the package versions to 0.10.0.
The sophomore release of vcd-ext-sdk packages is imminent!

In addition to updating the version number, this change removes the
dry-run status from NPM packaging, and re-enables the deploy filters for
Travis to only attempt a deploy on tagged releases from the master
branch.

Testing Done:
The release management angle of this was tested in the 06ec8ab commit,
where the dry-run status for NPM was verified in Travis, and the deploy
status for Nexus was verfied in, and dropped from, the Sonatype staging
repository.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>