Allow plugins to be written in Angular 6+.
OpaqueToken was removed as an Angular construct in Angular 6. The
container hooks used OpaqueToken so plugins that tried to use newer
versions of Angular and Clarity would result in compilation errors.

This change replaces OpaqueToken use with InjectionToken. Since
InjectionToken has been available since Angular 4, compatibility back to
vCD 9.1.0.2 was tested and successfully verified. While testing backward
compatibility, an issue with actAs functionality was discovered due to
the lack of a SESSION_ORG_ID token in a 9.1 environment. The SDK was
modified to log a warning instead of failing, when actAs can't be
automatically managed.

Testing Done:
* Consumed the new SDK from a 9.1-compatible seed project (Clarity
0.10.x and Angular 4.4.x). Verified successful compilation and
deployment to a vCD 9.1.0.2 testbed.
* Updated to package.json of seed plugin to utilize Angular 6 (also
required updating rxjs references and adding rxjs-compat dependency, and
@ngrx dependencies). Verified successful compilation of a plugin with
Angular 6. Deployed plugin to a 9.1.0.2 testbed and verified successful
loading of plugin.
* Updated Clarity to 0.13.x on top of the previous step. Verified
successful compilation and deployment to a vCD 10.0 environment.
* Made wild assumtion that 9.7 will also work.

Additional Testing Notes:
Peer dependency warning will be output for a plugin that wants to use
Angular 6+. These will be addressed in a patch version of the SDK.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>