Add Github Action for Typescript CI. (#120)
This change introduces a new YAML to define the build process for
Typescript projects. This is intended to be a replacement for Travis
integration.

The idea is to run this action on all pull requests and pushes to master
that contain Typescript changes. A future revision will add publishing
capabilities to this action.

A CI-specific karma configuration was added to use a headless
environment and to exit upon completion. This allows the existing `ng
test` to work in an ongoing way for local development.

Testing done:
- Ran `npm test` locally, verified test execution using Chrome
- Ran `npm run test-ci` locally, verified test execution and completion
using headless Chrome.

Signed-off-by: Jeff Moroski <jmoroski@vmware.com>