Add header links to response body.
This change normalizes the link processing to some extent
by moving the header-based links for /cloudapi endpoints
into the response body. The links can then be returned in the
payload to consumers of the VcdApiClient. The header links
were not previously accessible because the raw response is
not available outside of VcdApiClient.

As a reulst of this change, paging rels now work for queries in
the /cloudapi space.

Testing Done:
- Verify customize portal plugin is able to receive the header links
in the object's body.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Co-authored-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>