Fix link interceptor to return the links when 204 is returned and normalize the base path
#1 Issue
The link header intereceptor was not handling the case when empty body
is returned but links in the headers are present. - Improve the interecptor to
preserve the links even when the body is empty.

#2 Issue
When the base path is set to "" the client was transformin the  requests
to relative to current domain, instead of doing that we're now
normalizing the base path so the users can make requests using
absolute url paths. When they want to support relative to  their
current domain they can set the base path to "/".
Example:
#1 Absolute URL Path
this.client.basePath = ""
this.client.get("https://google.com")
Actual URL called: https://google.com

#2 Relative to current domain (http://localhost:4200)
this.client.basePath = "/"
this.client.get("my/request")
Actual URL called: http://localhost:4200/my/request

#3 Using hardcoded base path
this.client.basePath = "https://server.com"
this.client.get("my/request")
Actual URL called: https://server.com/my/request

Tesing Done:
- Verify that when request with 204 response but with links in the headers
is sent the result will contain the links.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>