Update the way how extension point schematic is configured
Create config file which will describe the schemtic for
certain type. The definision holds template name and steps
which has to be pefromed in form of functions.

Fix typos in schema.json of the extension point schematic.

Delete all .d.ts files which are actually not needed
to persist in the version control. They will be generated at
build time.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>