Add schematic for generating UI Extension Point (#146)
* Add schematic for generating UI Extension Point

Implement schematic for extension point generation.
The feature supports the following:
- Generate ext point of type entity action, i.e vm/vapp-action type
- Generate ext point of type primary nav
- Generate ext point of type view only, i.e navigation:application,
navigation:datacenter:* etc.

The schmatic is capable to:
- Gerate basic sources
- Import components or modules in your main module
- Register the extensions in the manifest

Testing Done:
- Generate all kinds of extension points
- Build -> Deploy -> Verifiied!

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

* Update the way how extension point schematic is configured

Create config file which will describe the schemtic for
certain type. The definision holds template name and steps
which has to be pefromed in form of functions.

Fix typos in schema.json of the extension point schematic.

Delete all .d.ts files which are actually not needed
to persist in the version control. They will be generated at
build time.

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>

Co-authored-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>