Plugin builder wasn't returning error code != 0 in case of an error
Issue: Plugin builder wasn't returning error code != 0 in case of an error
Fix: In case angular builder result is null return rejected promise,
pointing the user that the issue is with the vcd plugin builder.
In case that the angular builder fail just return the result
of the original angular builder.

Testing Done:
Verified after the fix the status code after failing build is != 0

Signed-off-by: Nikola Vladimirov Iliev <nvladimirovi@vmware.com>